add images updated

+++ service/household.py
@@ -30,13 +30,15 @@ class HouseholdService:
 
     @classmethod
     def add_image(cls, household_id: int, image_list: list, key: str, name: str):
-        for item in image_list:
-            code = base64.b64encode(item)
-            image = HouseHoldImage(image=code, household_id=household_id, created_at=datetime.datetime.now())
-            conn.add(image)
-        shutil.rmtree(f"images/{name}/{key}", ignore_errors=True)
-        conn.commit()
-        return True
+        if image_list:
+            for item in image_list:
+                code = base64.b64encode(item)
+                image = HouseHoldImage(image=code, household_id=household_id, created_at=datetime.datetime.now())
+                conn.add(image)
+            shutil.rmtree(f"images/{name}/{key}", ignore_errors=True)
+            conn.commit()
+            return True
+        return
 
     @classmethod
     def get_image(cls, key: str, household_id: int):
